Separate `store()` bill logic from main `BillsController`
